강학생 신청한것만 보이는 화면

+++ src/main/java/com/green/greenuni/application/lectures/model/LectureCreateReq.java
@@ -30,6 +30,6 @@ public class LectureCreateReq {
     private String dayOfWeek;
     private Integer startPeriod;
     private Integer endPeriod;
-    private Integer lectureId;
+    private Long lectureId;
 
 }
